requirements.txt
scrapy==2.0.1
tweepy==3.8.0
jsonlines==1.2.0
Twisted==20.3.0
pyenchant==3.0.1
nltk==3.5
textblob==0.15.3
supervisor==4.1.0
kafka-python==2.0.1
elasticsearch==7.6.0
pymongo==3.10.1